requirements.txt
requests
pyjwt
pytz
pyqrcode
pypng